fix(GoogleAnalytics): change structure of GA
- Add GoogleAnalytics component
- Add googleAnalytics util to enable it
