+++ .eslintrc.json
@@ -10,6 +10,7 @@
     "@typescript-eslint/no-explicit-any": "warn",
     "@typescript-eslint/require-await": "off",
     "@typescript-eslint/ban-ts-comment": "off",
-    "@typescript-eslint/no-empty-object-type": "warn"
+    "@typescript-eslint/no-empty-object-type": "warn",
+    "prefer-rest-params": "off"
   }
 }
