+++ src/app/components/GoogleAnalytics/index.tsx
@@ -1,0 +1,19 @@
+import Script from "next/script"
+
+export const GoogleAnalytics = () => {
+  return (
+    <>
+      <Script
+        src={`https://www.googletagmanager.com/gtag/js?id=${process.env.NEXT_PUBLIC_GOOGLE_ANALYTICS_ID}`}
+        strategy="afterInteractive"
+      />
+      <Script id="ga-setup" strategy="afterInteractive">
+        {`
+              window.dataLayer = window.dataLayer || [];
+              function gtag(){dataLayer.push(arguments);}
+              gtag('js', new Date());
+              `}
+      </Script>
+    </>
+  )
+}
